add dialog logic

+++ crossFit-app/src/app/workout/details-workout/details-workout.component.ts
@@ -3,6 +3,8 @@ import { WorkoutService } from '../workout.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Workout } from 'src/app/interfaces/workout';
 import { UserService } from 'src/app/user/user.service';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteConfirmationDialogComponent } from '../delete-confirm-dialog/delete-confirm-dialog.component';
 
 @Component({
   selector: 'app-details-workout',
@@ -27,6 +29,7 @@ export class DetailsWorkoutComponent implements OnInit {
     private workoutService: WorkoutService,
     private userService: UserService,
     private activatedRoute: ActivatedRoute,
+    private dialog: MatDialog,
     private router: Router) { }
 
   wokroutId = this.activatedRoute.snapshot.params?.['workoutId'] // така взимам ID-то на конкретна тренировка;
@@ -47,15 +50,43 @@ export class DetailsWorkoutComponent implements OnInit {
 
     if (this.currentWorkout) {
 
-      this.workoutService.deleteWorkout(this.wokroutId).subscribe(() => {
+      const dialogRef = this.dialog.open(DeleteConfirmationDialogComponent, {
+        data: {
+          title: 'Delete this workout',
+          message: 'Are you sure you want to delete this workout?',
+        },
+      });
 
-        this.router.navigateByUrl('/workouts/allWorouts');
-        console.log('Workout deleted successfully');
-      },
-        (error) => {
-          console.log('Error deleting workout', error);
+      dialogRef.afterClosed().subscribe((result) => {
+
+        if (result) {
+          this.workoutService.deleteWorkout(this.wokroutId).subscribe(() => {
+
+            this.router.navigateByUrl('/workouts/allWorouts');
+            console.log('Workout deleted successfully');
+          },
+            (error) => {
+              console.log('Error deleting workout', error);
+            }
+          )
         }
-      )
+
+      })
+
+      // const shouldDelete = window.confirm('Are you sue you want to delete this wokrout');
+
+      // if (shouldDelete) {
+
+      // this.workoutService.deleteWorkout(this.wokroutId).subscribe(() => {
+
+      //   this.router.navigateByUrl('/workouts/allWorouts');
+      //   console.log('Workout deleted successfully');
+      // },
+      //   (error) => {
+      //     console.log('Error deleting workout', error);
+      //   }
+      // )
+      // }
     }
   }
